chore: migrate splashtinder
